UBUNTU: SAUCE: apparmor4.0.0 [8/91]: Audit: Update shutdown LSM data
The audit process LSM information is changed from a secid audit_sig_sid
to an lsmblob in audit_sig_lsm. Update the users of this data
appropriately. Calls to security_secid_to_secctx() are changed to use
security_lsmblob_to_secctx() instead. security_current_getsecid_subj()
is scaffolded. It will be updated in a subsequent patch.

Signed-off-by: Casey Schaufler <casey@schaufler-ca.com>
Signed-off-by: Timo Aaltonen <timo.aaltonen@canonical.com>